+++ src/event.js
@@ -66,18 +66,20 @@ export function activeLink (dom, activeParent) {
   if (!dom) return
   let target
 
-  ;[].slice.call(dom.querySelectorAll('a')).forEach(node => {
-    if (node.href === host && !target) {
-      activeParent
-        ? node.parentNode.setAttribute('class', 'active')
-        : node.setAttribute('class', 'active')
-      target = node
-    } else {
-      activeParent
-        ? node.parentNode.removeAttribute('class')
-        : node.removeAttribute('class')
-    }
-  })
+  ;[].slice.call(dom.querySelectorAll('a'))
+    .sort((a, b) => b.href.length - a.href.length)
+    .forEach(node => {
+      if (host.indexOf(node.href) === 0 && !target) {
+        activeParent
+          ? node.parentNode.setAttribute('class', 'active')
+          : node.setAttribute('class', 'active')
+        target = node
+      } else {
+        activeParent
+          ? node.parentNode.removeAttribute('class')
+          : node.removeAttribute('class')
+      }
+    })
 
   return target
 }
@@ -92,19 +94,14 @@ export function bindToggle (dom) {
 
   dom.addEventListener('click', () => body.classList.toggle('close'))
 
-  if (!/mobile/i.test(navigator.userAgent)) return
-  document.querySelector('aside').addEventListener('click', event => {
-    body.classList.toggle('close')
-  })
+  if (isMobile()) {
+    document.querySelector('aside')
+      .addEventListener('click', _ => body.classList.toggle('close'))
+  }
 }
 
-let cacheContentDOM
 export function scroll2Top () {
-  if (!cacheContentDOM) {
-    const dom = isMobile() ? 'body' : 'section.content'
-    cacheContentDOM = document.querySelector(dom)
-  }
-  cacheContentDOM.scrollTop = 0
+  document.body.scrollTop = 0
 }
 
 export function sticky () {